// src/app/models/panel/admin/product/add-product.ts
export interface AddProduct {
    productName: string;
    productPrice: number;
    productCount: number;
    isEnable: boolean;
    size: string;
    code: string;
    genderOfConsumer: number;
    enclosureType: string;
    manufacturingCountry: string;
    manufacturerCompany: string;
    webAddress: string;
    features: string;
    methodUse: string;
    indications: string;
    warnings: string;
    maintenance: string;
    isSpecial: boolean;
    scientificName: string;
    subMenuId: string;
}